Upgrade to Kotlin 1.2.51 and serialization 0.6.0, thereby fixing 2 bugs as anticipated
Upgrading broke some code in `UnknownPolymorphicSerializer` which now needs to rely on a bit more reflection since the serialization 0.6.0 library closed of an internal type which was used.

In addition, all 'default' types will need to apply custom serializers where unknown types are expected on some of its containing objects. This is needed for deserialization only, thus this is less important for extending external types.