Do not use `DeploymentStatus` as snapshot, and add exception for non-deployable studies.
We will have to wait and see how to deal with persistence of `Deployment`. Since this does not need to be extensible, it would be okay to store this in a relational database (some ORM would be needed then). However, it is a bit unclear what the benefit would be.