Participant disambiguation (#96)
The previously implemented `UserService` of the study subsystem was an ambiguous mess of what we wanted to do with participants through the study manager, and how participants are linked to study deployments in the deployment subsystem.

This commit removes the notion of `Participant` from the study subsystem, and introduces the notion of a `Participation` in the deployment subsystem. Once a `Participation` is created, an `AccountService` is used to invite/create accounts and send out invitations.

This prepares us to implement a distinct `Participant` in the study subsystem, which is a placeholder for the researcher to add participant information, prior to it being linked to any deployment.

Commit log:

* Refactor: move `UserService` to deployment subsystem.

* Refactor: move user domain classes to `carp.common`.

* Replace 'participant' endpoints with 'participation' endpoints.

* Move account stuff from `UserService` to `carp.common` `AccountService`.

* Refactor: rename deployment `UserService` to `ParticipationService`.

* Refactor: rename deployment `UserRepository` to `ParticipationRepository`.

* Merge `AccountRepository` into `AccountService`.

* Incorporate study/account invitation fully into `AccountService`.

* Refactor: move `AccountService` back to deployment.

The methods in `AccountService` will need to take parameters which are specific to the deployment subsystem, i.e., study deployment information. Therefore, this cannot be in common. Likely, each different subsystem that has to do account operations needs to implement those in a separate `AccountService`.

* Bugfix: using 'else if' returned linter errors.

* Include `StudyInvitation` in participation invites.

* Merge `ParticipationService` into `DeploymentService`.

* Remove unused `DeploymentService.getParticipationsForStudyDeployment` use case.

* Bugfix: do not allow adding participations for non-existing deployments.

* Store study invitiation within participation.

Before, the invitation was only used as part of the invite sent through `AccountService`. Now, the invitation is also logged/stored.