Replace `MasterDeviceDeployment` checksum with `lastUpdateDate`. (#138)
* Replace `MasterDeviceDeployment` checksum with `lastUpdateDate`.

This is a more robust implementation to verify on `deviceDeployed` whether the most recent `MasterDeviceDeployment` was deployed.

However, on the JS runtime, the timer is sometimes not precise enough to return a new `DateTime` in between two device registrations. Therefore, I added an artificial wait period in the unit test verifying invalidated device deployments. In a real-world distributed environment, this should not happen.