Upgrade to Kotlin 1.4.0, kotlinx.serialization 1.0.0-RC, and Gradle 6.6.
This currently makes `verifyTsDeclarations` fail since the new JavaScript IR backend obfuscates types which are not exported using `@JsExport`. But, this currently does not work due to a bug in `kotlinx.serialization`: https://github.com/Kotlin/kotlinx.serialization/issues/980

We should also verify whether we want to set `allowStructuredMapKeys` to true in `createDefaultJSON()`, which seemingly was done automatically in the past but now is opt-in.