Switch to JS LEGACY backend.
It seems like the IR backend requires `@JsExport` to be applied since otherwise output names are heavily modified. But, `JsExport` does not seem that mature yet, so we will try to stick with the LEGACY backend for now.

There is a bug when running JS unit tests for serializers added to custom-named companion objects, so I had to remove the `PhoneSensorMeasure.Factory` name.