Rewrite `UnknownPolymorphicSerializer` to be registered as `default`.
`default` in polymorphic serialization is a feature in `kotlinx.serialization` which has been introduced after `UnkownPolymorphicSerializer` was first implemented. By using this new feature, it is no longer necessary to define the custom serializer everywhere the polymorphic type is expected. The serializer will only be triggered if an unknown type is encountered.

Fixes #9
Fixes #161

However, with the current implementation wrapper objects can only be serialized polymorphically, and not as a top-level object. This makes one test still fail in this commit.