Register `UnknownPolymorphicSerializer`s as `default` in `SerializersModule`. (#232)
By relying on the `default` fallback mechanism of polymorphic serialization to register the `UnknownPolymorphicSerializer`, it is no longer needed to specify the unknown serializers anywhere unknown types can be encountered in the codebase.

To make this work, the wrapper objects used to store unknown types at runtime also need to be registered as `subclass` for the polymorphic types. I updated the documentation accordingly.

In addition, this means that by registering a different `default` serializer (and `subclass` for the wrapper), unknown polymorphic serialization can now also be supported for different serialization formats than Json.

Making this work was a bit of a hack which required accessing `kotlinx.serialization` internals. An `Encoder` maintains state that the current object is being encoded as polymorphic or not. Depending on this state, `beginStructure` outputs type information or not. But, given that the wrapped object already contains type information, we never want to output this. But whether or not the `Encoder` is currently writing something as polymorphic is handled outside of the serializer. I wrote a tiny multiplatform accessor to modify this internal boolean state prior to encoding the wrapped json object. I will contact the Jetbrains folks to see whether or not they could consider this as a valid use case to open up their APIs a bit.

Fixes #161 
Fixes #9