Make `Measure` a sealed class with subclasses `DataStream` and `TriggerData`.
`TriggerData` can't be used yet in the domain model since `StudyProtocolSnapshot` needs to be serialized before `triggerId` is known which is needed in the constructor. Therefore, a `Measure` domain object needs to be introduced next, rather than reusing the current `Measure` DTO in the API, which can hold a `Trigger` reference until serialization happens.