Add endpoints to open/close data streams to `DataStreamService`.
Since all `DataStreamService` calls were essentially simple redirects to `DataStreamRepository`, I removed the intermediate repo. This is in line with a data ingestion service not mapping as nicely to the DDD repo pattern as in the other application services. Our current thinking is to avoid loading domain objects when ingesting data, leaving it open to specific infrastructure how the service constraints are implemented.