Move `DeviceDescriptor.getDefaultSamplingConfiguration` to client subsystem in `RuntimeDeviceInfo`.
This was originally added to `common` to support the client. This caused confusion in terms of overloaded names. By considering the use case why this needs to be called in the client subsystem, it turned out this method call likely will become irrelevant. The goal of `DataListener` is to communicate required sampling configurations to data collectors, and whether or not a passed sampling configuration is a "default" one or one stemming from a measure will become a hidden implementation detail. To support that implementation detail, this information is now made available as part of `RuntimeDeviceInfo` which can be obtained through `MasterDeviceDeployment`. This latter class is intended to be encapsulated by the client API in the end.