Fix: at least one device needs to be deployed in protocol with only optional devices.
In addition, in device status it is more meaningful to communicate whether a device can be deployed, otherwise information is lost. Whether it is required to be deployed can be inferred through the status of other devices and whether there are `remainingDevicesToRegisterToObtainDeployment`. However, we could consider also adding a `requiresDeployment` field to make this clearer.