`IntegrationEvent.aggregateId` informs infra about ordering requirements.
To make this work, `DeploymentService.Event.StudyDeploymentsRemoved` had to be changed to report on the removal of a single deployment. Otherwise there was no way to guarantee correct partitioning on event bus implementations.

We considered the downside of having more repository hits (one per delete), but considered that this could easily be batched by the event consumer in the infrastructure if needed.