Fix: role and expected participant data storage in protocol.
As part of this, I removed the ability to remove roles. This is unnecessary for now, and has side effects that need to be communicated as `StudyProtocol` events. To make this easy this needs to be refactored in the future, but can be avoided by removing this function for now.