Passing TS unit tests for kotlin/serialization/datetime/carp.common
Unit tests previously present for deployments, protocols, and studies subsystem have not been added yet. For now, this does outline a feasible project setup to start using a combination of generated TS declarations and manually added ones. Actually ensuring all necessary types are present will be postponed for later, for one, once upgrading to Kotlin 1.6.20 is possible, since it can generate declarations for enums and interfaces which Kotlin 1.6.10 we are currently stuck on cant'.