Support measurement validation on more generic maps.
This indirectly closes #389. For validation purposes, multiple `EnumObjectMap`s can easily be merged since that operation is supported on the base `Map` interface.

This was preferred over adding a `addAll` method on `EnumObjectMap` or extending classes, since it would only partially comply with the intent of `add` of these maps. The point is that not only `add` is called, but also that the result is stored in a local field to mimic an enum. This is not possible without manually constructing a new map which adds and assigned individual fields of other maps. But, this is not needed if the only use case is validation!