Refactor: don't pass deployment status to `ParticipantGroup`
This removes the dependency of `ParticipantGroup` on `StudyDeploymentStatus`, which I'm not certain was bad, but currently went unused. It prevented being able to call `markAsInvited` (now renamed to `markAsDeployed`) before having called `createStudyDeployment`. This was in fact by design.

However, it complicated handling concurrent calls to `RecruitmentServiceHost.inviteNewParticipantGroup` which we currently need to support. Relying on #398, we can detect concurrent calls and prevent calling `createStudyDeployment` in case of concurrent edits to the `Recruitment`, so that only successful `inviteNewParticipantGroup` calls would trigger actual study deployment creation.

But, this thus requires being able to mark a `ParticipantGroup` as deployed before `createStudyDeployment` is called.

Ideally, this should be implemented using eventual consistency instead (#295). Or, in fact, creating and inviting participant groups should really be separate calls, and `inviteNewParticipantGroup` should be made obsolete.

Until then (2.0 milestone most likely), this will at least resolve current concurrency issues.