Add EDA and PPG to Data JSON schema; include in checklist
When adding the new EDA measure type before, I forgot to add the EDA JSON schema to the Data JSON schema. This meant that EDA data wouldn't be validated (always be presumed correct).

I fixed this for EDA and PPG data, and included it in the development checklist so it's less likely this will be forgotten in the future.