Use JS IR backend; add Kotlin stdlib TypeScript facade
All TypeScript declarations and unit tests were removed since they need to be rewritten to match the new JS backend. This is already done for the Kotlin stdlib. But, to make this work on the IR backend, a facade for the Kotlin standard library is included, along with a build setup to hack into the generated Javascript sources to make this work.