Fix #437: incompatible TS types of duplicated types across modules
When using TS declarations, a cast to 'any' was needed for some types when a type from one module was passed to another module; this was observed with `StudyDeploymentStatus` (deployment subsystem) which was passed to `ParticipantGroupStatus` (studies subsystem).

This change fixes that by only loading generated TS declarations of a single generated module, which automatically includes dependent declarations. But, this required importing the dependent modules in the corresponding namespaces in the facade module; otherwise this is "undefined" at runtime.